Moved routes to own module.
