+++ recw/src/main.rs
@@ -1,36 +1,19 @@
-extern crate rustc_serialize;
 #[macro_use] extern crate nickel;
-
-use nickel::{Nickel, HttpRouter, Request, Response, MediaType, MiddlewareResult};
+use nickel::{Nickel, HttpRouter};
 use nickel::hyper::method::Method;
 
-use rustc_serialize::json;
+mod routes;
 
-#[derive(RustcDecodable, RustcEncodable)]
-struct Person {
-    firstname: String,
-    lastname:  String,
-}
-
-fn get_person<'r, 'mw, 'conn>(_: &mut Request<'mw, 'conn>, response: Response<'mw>) -> MiddlewareResult<'mw> {
-    let person = Person{
-        firstname: String::from("Bob"),
-        lastname: String::from("Dylan"),
-    };
-    let json = json::encode(&person).unwrap();
-    response.send(json)
-}
-
-fn hello(word: &str) -> String
-{
-    String::from(format!("Hello {}", word))
-}
 
 fn main() {
     let mut server = Nickel::new();
 
-    server.get("/", middleware!({format!("{}", hello("index"))}));
-    server.get("/hello", middleware!({format!("{}", hello("world"))}));
-    server.add_route(Method::Get, "/person", get_person);
-    server.listen("127.0.0.1:6767");
+    server.get("/", middleware!({format!("{}", routes::hello("index"))}));
+    server.get("/hello", middleware!({format!("{}", routes::hello("world"))}));
+    server.add_route(Method::Get, "/person", routes::get_person);
+    match server.listen("127.0.0.1:6767") {
+        Ok(_res) => {println!("Server has started")}
+        Err(e) => {println!("Server failed to start: {:?}", e)}
+    }
 }
+
